+++ src/js/slider.js
@@ -27,6 +27,10 @@ vjs.Slider = vjs.Component.extend({
     player.on(this.playerEvent, vjs.bind(this, this.update));
 
     this.boundEvents = {};
+
+
+    this.boundEvents.move = vjs.bind(this, this.onMouseMove);
+    this.boundEvents.end = vjs.bind(this, this.onMouseUp);
   }
 });
 
@@ -48,9 +52,7 @@ vjs.Slider.prototype.createEl = function(type, props) {
 vjs.Slider.prototype.onMouseDown = function(event){
   event.preventDefault();
   vjs.blockTextSelection();
-
-  this.boundEvents.move = vjs.bind(this, this.onMouseMove);
-  this.boundEvents.end = vjs.bind(this, this.onMouseUp);
+  this.addClass('vjs-sliding');
 
   vjs.on(document, 'mousemove', this.boundEvents.move);
   vjs.on(document, 'mouseup', this.boundEvents.end);
@@ -60,8 +62,13 @@ vjs.Slider.prototype.onMouseDown = function(event){
   this.onMouseMove(event);
 };
 
+// To be overridden by a subclass
+vjs.Slider.prototype.onMouseMove = function(){};
+
 vjs.Slider.prototype.onMouseUp = function() {
   vjs.unblockTextSelection();
+  this.removeClass('vjs-sliding');
+
   vjs.off(document, 'mousemove', this.boundEvents.move, false);
   vjs.off(document, 'mouseup', this.boundEvents.end, false);
   vjs.off(document, 'touchmove', this.boundEvents.move, false);
@@ -117,7 +124,9 @@ vjs.Slider.prototype.update = function(){
   }
 
   // Set the new bar width
-  bar.el().style.width = vjs.round(barProgress * 100, 2) + '%';
+  if (bar) {
+    bar.el().style.width = vjs.round(barProgress * 100, 2) + '%';
+  }
 };
 
 vjs.Slider.prototype.calculateDistance = function(event){
@@ -128,7 +137,7 @@ vjs.Slider.prototype.calculateDistance = function(event){
   boxW = boxH = el.offsetWidth;
   handle = this.handle;
 
-  if (this.options_.vertical) {
+  if (this.options()['vertical']) {
     boxY = box.top;
 
     if (event.changedTouches) {
@@ -174,10 +183,10 @@ vjs.Slider.prototype.onFocus = function(){
 };
 
 vjs.Slider.prototype.onKeyPress = function(event){
-  if (event.which == 37) { // Left Arrow
+  if (event.which == 37 || event.which == 40) { // Left and Down Arrows
     event.preventDefault();
     this.stepBack();
-  } else if (event.which == 39) { // Right Arrow
+  } else if (event.which == 38 || event.which == 39) { // Up and Right Arrows
     event.preventDefault();
     this.stepForward();
   }